+++ include/Module/CRC/Polynomial/CRC_polynomial_inter.hpp
@@ -26,10 +26,10 @@ public:
 	virtual CRC_polynomial_inter<B>* clone() const;
 
 protected:
-	virtual void _build       (const B *U_K1, B *U_K2, const int frame_id);
-	virtual bool _check       (const B *V_K          , const int frame_id);
-	virtual bool _check_packed(const B *V_K          , const int frame_id);
-	virtual void _extract     (const B *V_K1, B *V_K2, const int frame_id);
+	virtual void _build       (const B *U_K1, B *U_K2, const size_t frame_id);
+	virtual bool _check       (const B *V_K          , const size_t frame_id);
+	virtual bool _check_packed(const B *V_K          , const size_t frame_id);
+	virtual void _extract     (const B *V_K1, B *V_K2, const size_t frame_id);
 
 	void _generate_INTER(const B *U_in,
 	                           B *U_out,
