+++ include/Module/Channel/Rayleigh/Channel_Rayleigh_LLR_user.hpp
@@ -52,10 +52,10 @@ public:
 
 	void set_seed(const int seed);
 
-	virtual void set_n_frames(const int n_frames);
+	virtual void set_n_frames(const size_t n_frames);
 
 protected:
-	virtual void _add_noise_wg(const R *X_N, R *H_N, R *Y_N, const int frame_id);
+	virtual void _add_noise_wg(const R *X_N, R *H_N, R *Y_N, const size_t frame_id);
 
 	virtual void deep_copy(const Channel_Rayleigh_LLR_user<R>& m);
