+++ include/Module/Decoder/BCH/Fast/Decoder_BCH_fast.hpp
@@ -42,15 +42,15 @@ public:
 	virtual ~Decoder_BCH_fast() = default;
 	virtual Decoder_BCH_fast<B,R>* clone() const;
 
-	virtual void set_n_frames(const int n_frames);
+	virtual void set_n_frames(const size_t n_frames);
 
 protected:
-	virtual int  _decode        (                      const int frame_id);
-	virtual void _load          (const B *Y_N,         const int frame_id);
-	virtual int  _decode_hiho   (const B *Y_N, B *V_K, const int frame_id);
-	virtual int  _decode_hiho_cw(const B *Y_N, B *V_N, const int frame_id);
-	virtual int  _decode_siho   (const R *Y_N, B *V_K, const int frame_id);
-	virtual int  _decode_siho_cw(const R *Y_N, B *V_N, const int frame_id);
+	virtual int  _decode        (                      const size_t frame_id);
+	virtual void _load          (const B *Y_N,         const size_t frame_id);
+	virtual int  _decode_hiho   (const B *Y_N, B *V_K, const size_t frame_id);
+	virtual int  _decode_hiho_cw(const B *Y_N, B *V_N, const size_t frame_id);
+	virtual int  _decode_siho   (const R *Y_N, B *V_K, const size_t frame_id);
+	virtual int  _decode_siho_cw(const R *Y_N, B *V_N, const size_t frame_id);
 };
 }
 }
