+++ include/Module/Decoder/Generic/ML/Decoder_maximum_likelihood_std.hpp
@@ -29,10 +29,10 @@ public:
 	virtual Decoder_maximum_likelihood_std<B,R>* clone() const;
 
 protected:
-	int _decode_siho   (const R *Y_N,  B *V_K, const int frame_id);
-	int _decode_siho_cw(const R *Y_N,  B *V_N, const int frame_id);
-	int _decode_hiho   (const B *Y_N,  B *V_K, const int frame_id);
-	int _decode_hiho_cw(const B *Y_N,  B *V_N, const int frame_id);
+	int _decode_siho   (const R *Y_N,  B *V_K, const size_t frame_id);
+	int _decode_siho_cw(const R *Y_N,  B *V_N, const size_t frame_id);
+	int _decode_hiho   (const B *Y_N,  B *V_K, const size_t frame_id);
+	int _decode_hiho_cw(const B *Y_N,  B *V_N, const size_t frame_id);
 };
 
 template <typename B = int, typename R = float>
