+++ include/Module/Decoder/LDPC/BF/Decoder_LDPC_bit_flipping_hard.hpp
@@ -42,19 +42,19 @@ public:
 	virtual Decoder_LDPC_bit_flipping_hard<B,R>* clone() const;
 
 protected:
-	void _store         (B *V_K,               const int frame_id);
-	void _store_cw      (B *V_N,               const int frame_id);
+	void _store        (B *V_K,               const size_t frame_id);
+	void _store_cw     (B *V_N,               const size_t frame_id);
 
-	int _decode_siho   (const R *Y_N, B *V_K, const int frame_id);
-	int _decode_siho_cw(const R *Y_N, B *V_N, const int frame_id);
+	int _decode_siho   (const R *Y_N, B *V_K, const size_t frame_id);
+	int _decode_siho_cw(const R *Y_N, B *V_N, const size_t frame_id);
 
-	int _decode_hiho   (const B *Y_N, B *V_K, const int frame_id);
-	int _decode_hiho_cw(const B *Y_N, B *V_N, const int frame_id);
+	int _decode_hiho   (const B *Y_N, B *V_K, const size_t frame_id);
+	int _decode_hiho_cw(const B *Y_N, B *V_N, const size_t frame_id);
 
 	// return true if the syndrome is valid
-	virtual bool     decode(const B *Y_N,                     const int frame_id);
-	virtual void cn_process(const B *VN,  B *CN,              const int frame_id) = 0;
-	virtual void vn_process(const B *Y_N, B *VN, const B *CN, const int frame_id) = 0;
+	virtual bool     decode(const B *Y_N,                     const size_t frame_id);
+	virtual void cn_process(const B *VN,  B *CN,              const size_t frame_id) = 0;
+	virtual void vn_process(const B *Y_N, B *VN, const B *CN, const size_t frame_id) = 0;
 
 	bool check_syndrome() const;
 };
